+++ src/hyp3_isce2/burst.py
@@ -6,6 +6,7 @@ import time
 from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass
 from pathlib import Path
+from secrets import token_hex
 from typing import Iterator, List, Optional, Tuple, Union
 
 import asf_search
@@ -327,28 +328,46 @@ def download_bursts(param_list: Iterator[BurstParams]) -> List[BurstMetadata]:
 def get_product_name(
     reference_scene: str,
     secondary_scene: str,
+    pixel_spacing: int
 ) -> str:
     """Get the name of the interferogram product.
 
     Args:
         reference_scene: The reference burst name.
         secondary_scene: The secondary burst name.
+        pixel_spacing: The spacing of the pixels in the output image.
 
     Returns:
         The name of the interferogram product.
     """
-    # If this changes, we will also need to update the burst product README template,
-    # which documents this naming convention.
-    return f'{reference_scene}x{secondary_scene}'
 
+    reference_split = reference_scene.split('_')
+    secondary_split = secondary_scene.split('_')
 
-def search_cmr_uat(scene_name: str) -> asf_search.ASFSearchResults:
-    opts = asf_search.ASFSearchOptions(host='cmr.uat.earthdata.nasa.gov')
-    return asf_search.search(product_list=[scene_name], opts=opts)
+    platform = reference_split[0]
+    burst_id = reference_split[1]
+    image_plus_swath = reference_split[2]
+    reference_date = reference_split[3][0:8]
+    secondary_date = secondary_split[3][0:8]
+    polarization = reference_split[4]
+    product_type = 'INT'
+    pixel_spacing = str(int(pixel_spacing))
+    product_id = token_hex(2).upper()
+
+    return '_'.join([
+        platform,
+        burst_id,
+        image_plus_swath,
+        reference_date,
+        secondary_date,
+        polarization,
+        product_type + pixel_spacing,
+        product_id
+    ])
 
 
 def get_burst_params(scene_name: str) -> BurstParams:
-    results = search_cmr_uat(scene_name)
+    results = asf_search.search(product_list=[scene_name])
 
     if len(results) == 0:
         raise ValueError(f'ASF Search failed to find {scene_name}.')